requirements.txt
asgiref==3.3.1
Django==3.0.9
djangorestframework==3.12.2
psycopg2==2.8.6
pytz==2021.1
sqlparse==0.4.1
ipdb==0.13.7
django-extensions==3.0.9
whitenoise==5.3.0
gunicorn==20.0.4
django-cors-headers==3.10.1
requests==2.27.1